+++ server/internal/app/models.go
@@ -7,7 +7,7 @@ import (
 	"time"
 )
 
-var Version = "1.5.0-rc14"
+var Version = "1.5.0-rc15"
 
 type SDRDevice struct {
 	ID                      string             `json:"id"`